Updated meta.json validation
Superwhisper will soon change the behavior when writing the result to the meta.json file. Duration is no longer a reliable validation test to check. Instead, it checks if there's an llmModel set and if so it will wait for llmResult. If not, it will only wait for the simple result.